model-server/requirements.txt
aiofiles==23.2.1
annotated-types==0.5.0
anyio==3.7.1
bchlib==1.0.0
beautifulsoup4==4.12.2
black==23.9.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudinary==1.34.0
distlib==0.3.7
fastapi==0.103.1
filelock==3.12.3
gdown==4.7.1
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
priority==2.0.0
psycopg2==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
PySocks==1.7.1
python-multipart==0.0.6
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
torch==2.0.1
torchgeometry==0.1.2
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
virtualenv==20.24.5
wsproto==1.2.0
